multiprocessing

+++ models/training.py
@@ -6,14 +6,15 @@ import os
 from torch.utils.tensorboard import SummaryWriter
 from glob import glob
 import numpy as np
-
+import torch.distributed as dist
 
 
 
 class Trainer(object):
 
-    def __init__(self, model, device, train_dataset, val_dataset, exp_name, optimizer='Adam'):
-        self.model = model.to(device)
+    def __init__(self, model, device, train_dataset, val_dataset, exp_name, rank, world_size, optimizer='Adam', parallel = False):
+        # self.model = model.to(device)
+        self.model = model
         self.device = device
         if optimizer == 'Adam':
             self.optimizer = optim.Adam(self.model.parameters(), lr=1e-4)
@@ -31,6 +32,9 @@ class Trainer(object):
             os.makedirs(self.checkpoint_path)
         self.writer = SummaryWriter(self.exp_path + 'summary'.format(exp_name))
         self.val_min = None
+        self.rank = rank
+        self.world_size = world_size
+        self.parallel = parallel
 
 
     def train_step(self,batch):
@@ -62,7 +66,9 @@ class Trainer(object):
 
     def train_model(self, epochs):
         loss = 0
-        start = self.load_checkpoint()
+        # start = self.load_checkpoint()
+        map_location = {'cuda:%d' % 0: 'cuda:%d' % self.rank}
+        start = self.load_checkpoint(map_location=map_location)
 
         for epoch in range(start, epochs):
             sum_loss = 0
@@ -70,20 +76,47 @@ class Trainer(object):
             train_data_loader = self.train_dataset.get_loader()
 
             if epoch % 1 == 0:
-                self.save_checkpoint(epoch)
-                val_loss = self.compute_val_loss()
+                if self.parallel :
+                    if not self.rank:
+                        self.save_checkpoint(epoch)
+                    dist.barrier()
+                    val_loss_from_others = torch.zeros(1)
+                    val_loss = self.compute_val_loss()
+                    if not self.rank:
+                        for rank in range(1,self.world_size):
+                            dist.recv(tensor=val_loss_from_others, src=rank)
+                            val_loss += val_loss_from_others
+                        val_loss = val_loss/self.world_size
+                    else:
+                        dist.send(tensor=val_loss, dst=0)
+                    dist.barrier()
+                    if not self.rank:
+                        if self.val_min is None:
+                            self.val_min = val_loss
 
-                if self.val_min is None:
-                    self.val_min = val_loss
-
-                if val_loss < self.val_min:
-                    self.val_min = val_loss
-                    for path in glob(self.exp_path + 'val_min=*'):
-                        os.remove(path)
-                    np.save(self.exp_path + 'val_min={}'.format(epoch),[epoch,val_loss])
+                            if val_loss < self.val_min:
+                                self.val_min = val_loss
+                                for path in glob(self.exp_path + 'val_min=*'):
+                                    os.remove(path)
+                                np.save(self.exp_path + 'val_min={}'.format(epoch),[epoch,val_loss])
 
 
-                self.writer.add_scalar('val loss batch avg', val_loss, epoch)
+                            self.writer.add_scalar('val loss batch avg', val_loss, epoch)
+                else:
+                    self.save_checkpoint(epoch)
+                    val_loss = self.compute_val_loss()
+
+                    if self.val_min is None:
+                        self.val_min = val_loss
+
+                    if val_loss < self.val_min:
+                        self.val_min = val_loss
+                        for path in glob(self.exp_path + 'val_min=*'):
+                            os.remove(path)
+                        np.save(self.exp_path + 'val_min={}'.format(epoch),[epoch,val_loss])
+
+
+                    self.writer.add_scalar('val loss batch avg', val_loss, epoch)
 
 
             for ib, batch in enumerate(train_data_loader):
@@ -103,7 +136,7 @@ class Trainer(object):
             torch.save({'epoch':epoch,'model_state_dict': self.model.state_dict(),
                         'optimizer_state_dict': self.optimizer.state_dict()}, path)
 
-    def load_checkpoint(self):
+    def load_checkpoint(self, map_location):
         checkpoints = glob(self.checkpoint_path+'/*')
         if len(checkpoints) == 0:
             print('No checkpoints found at {}'.format(self.checkpoint_path))
@@ -115,7 +148,7 @@ class Trainer(object):
         path = self.checkpoint_path + 'checkpoint_epoch_{}.tar'.format(checkpoints[-1])
 
         print('Loaded checkpoint from: {}'.format(path))
-        checkpoint = torch.load(path)
+        checkpoint = torch.load(path, map_location=map_location)
         self.model.load_state_dict(checkpoint['model_state_dict'])
         self.optimizer.load_state_dict(checkpoint['optimizer_state_dict'])
         epoch = checkpoint['epoch']
